Moved logger to own file
