+++ scripts/create_release.py
@@ -1,4 +1,5 @@
 from amalgamate import amalgamate
+from logger import logger
 
 import argparse
 from pathlib import Path
@@ -6,21 +7,6 @@ import shutil
 import re
 import zipfile
 import os
-
-
-class Logger:
-    def __init__(self, verbose=False):
-        self.verbose = verbose
-
-    def log(self, msg: str):
-        print(msg)
-
-    def verbose_log(self, msg: str):
-        if self.verbose:
-            print(msg)
-
-
-logger = Logger()
 
 
 def validate_version(version: str):
